Split the damage editor out of one class
UnitEditorDialog had grown to just over 2000 lines: it held every
control, built them for each unit type, laid them out, drew the armor
diagram, and wrote everything back onto the unit.

The work is now split by what it does, with no change in behavior:

- UnitDamageControls holds the controls, so the class that builds them
  and the class that reads them back share one set.
- UnitDamagePanelBuilder builds them, a builder per unit type, since
  every type carries its systems differently.
- UnitDamageApplier writes them back onto the unit, which can now be
  tested without a dialog.
- CheckCritPanel is its own class rather than an inner one.

UnitEditorDialog is left with the dialog: layout, the armor diagram,
the pre-existing damage roller, and the buttons. It drops from 2020
lines to 840.
